menambahkan if session pada setiap function admin

+++ application/controllers/admin/Ships.php
@@ -16,33 +16,53 @@ class Ships extends CI_Controller
 
 	public function index()
 	{
-		$data["allShips"] = $this->MShips->getAll();
-		$this->load->view("admin/ship/v_ListShips", $data);
+		if ($this->session->userdata('isloggedin')) {
+			$data["allShips"] = $this->MShips->getAll();
+			$this->load->view("admin/ship/v_ListShips", $data);
+		}else{
+			redirect('admin/user/login');
+		}
+
+		
 	}
 
 	public function add()
 	{
+		if ($this->session->userdata('isloggedin')) {
+			$this->load->view("admin/ship/v_AddShip");
+		}else{
+			redirect('admin/user/login');
+		}
+
 		if (isset($_POST['submit_ship'])) {
 			$this->MShips->save($_POST);
 			redirect("admin/ships");
 		}
-		$this->load->view("admin/ship/v_AddShip");
 	}
 
 	public function update($id)
 	{
+		if ($this->session->userdata('isloggedin')) {
+			$data["editShip"] = $this->MShips->getById($id);
+			$this->load->view("admin/ship/v_EditShip", $data);
+		}else{
+			redirect('admin/user/login');
+		}
+
 		if (isset($_POST['edit_ship'])) {
 			$this->MShips->update($_POST, $id);
 			redirect("admin/ships");
-		}
-		$data["editShip"] = $this->MShips->getById($id);
-		$this->load->view("admin/ship/v_EditShip", $data);
+		}		
 	}
 
 	public function delete($id)
 	{
-		$this->MShips->delete($id);
-		redirect("admin/ships");
+		if ($this->session->userdata('isloggedin')) {
+			$this->MShips->delete($id);
+			redirect("admin/ships");
+		}else{
+			redirect('admin/user/login');
+		}		
 	}
 }
 ?>
